- Modified VPC data resource on Terraform due to error when testing - Modified cluster name from var.addon_context.eks_cluster_id to var.eks_cluster_id - Modified entire three module guide and its written format. Made minor changes to commands so output is more efficient and less verbose. - had to update ssm.sh script path on the guide to absolute path - fixed cleanup script to include deleting new-cni-nodegroup - identified issue where aws-node pod was not in pending state. Source issue was the node type. Modified nodetype for the new manage nodegroup from m5.large to c5.large.

+++ manifests/modules/troubleshooting/cni/.workshop/cleanup.sh
@@ -19,12 +19,63 @@ if [ $is_policy_exist -eq 0 ]; then
     aws iam attach-role-policy --role-name $VPC_CNI_IAM_ROLE_NAME --policy-arn arn:aws:iam::aws:policy/AmazonEKS_CNI_Policy
 fi
 
+# Check if new-cni-nodes nodegroup exists and delete it specifically
+if aws eks describe-nodegroup --cluster-name $EKS_CLUSTER_NAME --nodegroup-name new-cni-nodes &>/dev/null; then
+    logmessage "Finding EC2 instances associated with new-cni-nodes nodegroup"
+    
+    # Get the Auto Scaling Group name associated with the nodegroup
+    ASG_NAME=$(aws eks describe-nodegroup \
+        --cluster-name $EKS_CLUSTER_NAME \
+        --nodegroup-name new-cni-nodes \
+        --query "nodegroup.resources.autoScalingGroups[0].name" \
+        --output text)
+    
+    if [ "$ASG_NAME" != "None" ] && [ -n "$ASG_NAME" ]; then
+        logmessage "Found Auto Scaling Group: $ASG_NAME"
+        
+        # Get instance IDs from the Auto Scaling Group
+        INSTANCE_IDS=$(aws autoscaling describe-auto-scaling-groups \
+            --auto-scaling-group-names $ASG_NAME \
+            --query "AutoScalingGroups[0].Instances[*].InstanceId" \
+            --output text)
+        
+        if [ -n "$INSTANCE_IDS" ]; then
+            logmessage "Found instances: $INSTANCE_IDS"
+            
+            # Terminate the instances asynchronously
+            logmessage "Terminating instances asynchronously"
+            aws ec2 terminate-instances --instance-ids $INSTANCE_IDS &
+        else
+            logmessage "No instances found in the Auto Scaling Group"
+        fi
+    else
+        logmessage "No Auto Scaling Group found for the nodegroup"
+    fi
+    
+    # Delete the nodegroup asynchronously
+    logmessage "Deleting nodegroup new-cni-nodes asynchronously"
+    aws eks delete-nodegroup --cluster-name $EKS_CLUSTER_NAME --nodegroup-name new-cni-nodes &
+    
+    # Store the PID of the delete-nodegroup process
+    NODEGROUP_DELETE_PID=$!
+    
+    # Wait for the nodegroup deletion to complete in the background
+    {
+        aws eks wait nodegroup-deleted --cluster-name $EKS_CLUSTER_NAME --nodegroup-name new-cni-nodes
+        logmessage "new-cni-nodes nodegroup deleted successfully"
+    } &
+    
+    # Continue with the rest of the script without waiting
+    logmessage "Continuing with other operations while nodegroup deletion happens in background"
+fi
+
+
 nodes=$(aws eks list-nodegroups --cluster-name $EKS_CLUSTER_NAME --query 'nodegroups' --output text)
 deleted_nodes=()
 
 logmessage "Reverting EKS managed nodegroup configuration"
 for node in ${nodes[@]}; do
-    if [[ "$node" != "default" && "$node" != "cni_troubleshooting_nodes" ]]; then
+    if [[ "$node" != "default" && "$node" != "cni_troubleshooting_nodes" && "$node" != "new-cni-nodes" ]]; then
         logmessage "Deleting nodegroup $node"
         aws eks delete-nodegroup --cluster-name $EKS_CLUSTER_NAME --nodegroup-name $node
         deleted_nodes+=$node
@@ -51,5 +102,6 @@ if [ $DEFAULT_CONFIG != $CURRENT_CONFIG ]; then
 
     aws eks update-addon --addon-name vpc-cni --cluster-name $EKS_CLUSTER_NAME --service-account-role-arn $VPC_CNI_IAM_ROLE_ARN --configuration-values $DEFAULT_CONFIG
 fi
+
